+++ RELEASE_NOTES.md
@@ -1,5 +1,12 @@
+### 0.4.2
+
+New features:
+ - Subplots common title
+ - Subplots margin
+ - Subplots common visibility
+
 ### 0.4.1 
 
-New feature:
+New features:
  - Tooltip delay control
  - External legend support in subplots